nullIfNotFound instead
