+++ src/errors.js
@@ -3,3 +3,12 @@ export class NotFoundError extends Error {
     super(message)
   }
 }
+
+/**
+ * @param {unknown} err
+ * @returns {null}
+ */
+export function nullIfNotFound(err) {
+  if (err instanceof NotFoundError) return null
+  throw err
+}
