+++ src/roles.js
@@ -2,7 +2,7 @@ import { currentSchemaVersions } from '@comapeo/schema'
 import mapObject from 'map-obj'
 import { kCreateWithDocId, kDataStore } from './datatype/index.js'
 import { assert, setHas } from './utils.js'
-import { getByDocIdIfExists } from './datatype/get-if-exists.js'
+import { nullIfNotFound } from './errors.js'
 import { TypedEmitter } from 'tiny-typed-emitter'
 /** @import { Namespace } from './types.js' */
 
@@ -270,7 +270,9 @@ export class Roles extends TypedEmitter {
    * @returns {Promise<Role>}
    */
   async getRole(deviceId) {
-    const roleRecord = await getByDocIdIfExists(this.#dataType, deviceId)
+    const roleRecord = await this.#dataType
+      .getByDocId(deviceId)
+      .catch(nullIfNotFound)
     if (!roleRecord) {
       // The project creator will have the creator role
       const authCoreId = await this.#coreOwnership.getCoreId(deviceId, 'auth')
@@ -385,7 +387,9 @@ export class Roles extends TypedEmitter {
       }
     }
 
-    const existingRoleDoc = await getByDocIdIfExists(this.#dataType, deviceId)
+    const existingRoleDoc = await this.#dataType
+      .getByDocId(deviceId)
+      .catch(nullIfNotFound)
 
     if (existingRoleDoc) {
       await this.#dataType.update(
